# docs/swagger.yaml
definitions:
  models.BuildParam:
    properties:
      arch:
        description: Id        int      `gorm:"primaryKey"`
        type: string
      buildType:
        type: string
      customPkg:
        items:
          type: string
        type: array
      desc:
        type: string
      label:
        type: string
      release:
        type: string
    type: object
  models.CreateUserInput:
    properties:
      address:
        type: string
      birthdate:
        type: string
      blocked:
        type: boolean
      browser:
        type: string
      company:
        type: string
      country:
        type: string
      device:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      externalId:
        type: string
      familyName:
        type: string
      formatted:
        type: string
      gender:
        type: string
      givenName:
        type: string
      isDeleted:
        type: boolean
      lastIP:
        type: string
      lastLogin:
        type: string
      locale:
        type: string
      locality:
        type: string
      loginsCount:
        type: integer
      middleName:
        type: string
      name:
        type: string
      nickname:
        type: string
      oauth:
        type: string
      openid:
        type: string
      password:
        type: string
      phone:
        type: string
      phoneVerified:
        type: boolean
      photo:
        type: string
      postalCode:
        type: string
      preferredUsername:
        type: string
      profile:
        type: string
      region:
        type: string
      registerSource:
        items:
          type: string
        type: array
      signedUp:
        type: string
      streetAddress:
        type: string
      unionid:
        type: string
      username:
        type: string
      website:
        type: string
      zoneinfo:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/auth/createUser:
    post:
      consumes:
      - application/json
      description: AuthingCreateUser
      parameters:
      - description: body for user info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserInput'
      produces:
      - application/json
      responses: {}
      summary: AuthingCreateUser
      tags:
      - Authing
  /v1/auth/getDetail/{authingUserId}:
    get:
      consumes:
      - application/json
      description: AuthingGetToken
      parameters:
      - description: The key for staticblock
        in: path
        name: authingUserId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: AuthingGetToken
      tags:
      - Authing
  /v1/auth/loginok:
    get:
      consumes:
      - application/json
      description: login success redirect url
      produces:
      - application/json
      responses: {}
      summary: login success redirect url
      tags:
      - Authing
  /v1/images/param/getBaseData/:
    get:
      consumes:
      - application/json
      description: get architecture, release Version, output Format ,and default package
        name list
      produces:
      - application/json
      responses: {}
      summary: GetBaseData param
      tags:
      - v1 job
  /v1/images/param/getCustomePkgList/:
    get:
      consumes:
      - application/json
      description: get custom package name list
      parameters:
      - description: ' arch ,e g:x86_64'
        in: query
        name: arch
        required: true
        type: string
      - description: 'release  '
        in: query
        name: release
        required: true
        type: string
      - description: 'custom group  '
        in: query
        name: sig
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: GetCustomePkgList param
      tags:
      - v1 job
  /v1/images/queryHistory/mine:
    get:
      consumes:
      - application/json
      description: Query My History
      parameters:
      - description: 'offset '
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: QueryMyHistory
      tags:
      - v1 job
  /v1/images/queryJobLogs/{name}:
    get:
      consumes:
      - application/json
      description: QueryJobLogs for given job name
      parameters:
      - description: The name for job
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: QueryJobLogs
      tags:
      - v1 job
  /v1/images/queryJobStatus/{name}:
    get:
      consumes:
      - application/json
      description: QueryJobStatus for given job name
      parameters:
      - description: The name for job
        in: path
        name: name
        required: true
        type: string
      - description: 'The id for job in database. '
        in: query
        name: id
        type: string
      - description: 'job namespace '
        in: query
        name: ns
        type: string
      produces:
      - application/json
      responses: {}
      summary: QueryJobStatus
      tags:
      - v1 job
  /v1/images/startBuild:
    post:
      consumes:
      - application/json
      description: start a image build job
      parameters:
      - description: body for ImageMeta content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.BuildParam'
      produces:
      - application/json
      responses: {}
      summary: StartBuild Job
      tags:
      - v1 job
  /v2/images/createJob:
    post:
      consumes:
      - application/json
      description: start a image build job
      parameters:
      - description: body for ImageMeta content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.BuildParam'
      produces:
      - application/json
      responses: {}
      summary: Create Job
      tags:
      - v2 job
  /v2/images/deleteJob/{id}:
    delete:
      consumes:
      - application/json
      description: delete a job build record
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: deleteRecord
      tags:
      - v2 job
  /v2/images/getJobParam/{id}:
    get:
      consumes:
      - application/json
      description: get job build param
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: GetJobParam
      tags:
      - v2 job
  /v2/images/getLogsOf/{id}:
    get:
      consumes:
      - application/json
      description: get single job logs
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      - description: job id
        in: query
        name: stepID
        required: true
        type: string
      - description: max Record number
        in: query
        name: maxRecord
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get single job logs
      tags:
      - v2 job
  /v2/images/getMySummary:
    get:
      consumes:
      - application/json
      description: get my summary
      produces:
      - application/json
      responses: {}
      summary: MySummary
      tags:
      - v2 job
  /v2/images/getOne/{id}:
    get:
      consumes:
      - application/json
      description: get single job detail
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get single job detail
      tags:
      - v2 job
swagger: "2.0"